requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6

pydantic==2.12.3
pydantic-settings==2.6.1
email-validator==2.3.0

SQLAlchemy==2.0.36
alembic==1.13.3
psycopg[binary]==3.2.3

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-multipart==0.0.12
pandas==2.2.3
